Optimize and clean up Vertex code

+++ GL/Texture.h
@@ -5,9 +5,9 @@
 #include <gl.h>
 
 struct Texture{
-	Texture(GLuint *texture = nullptr, Rect<float> rect = Rect<float>(0.f,0.f,0.f,0.f)) : m_texture(texture), m_rect(rect){}
+	Texture(GLuint *texture = nullptr, Rect<uint16_t> rect = Rect<uint16_t>(0,0,0,0)) : m_texture(texture), m_rect(rect){}
 	GLuint* m_texture;
-	Rect<float> m_rect;
+	Rect<uint16_t> m_rect;
 	uint16_t width=0;
 	uint16_t height=0;
 	char rotated = 0;
